examples renamed
