+++ examples/tsp/composing_criteria/composition_3criteria.rs
