+++ examples/tsp/composing_criteria/composition_5criteria.rs
