+++ examples/tsp/composing_criteria/explicit_3criteria.rs
