+++ examples/tsp/composing_criteria/explicit_5criteria.rs
