+++ examples/tsp/composing_criteria/mod.rs
@@ -1,5 +1,5 @@
-pub mod composition3;
-pub mod composition5;
-pub mod explicit1;
-pub mod explicit3;
-pub mod explicit5;
+pub mod composition_3criteria;
+pub mod composition_5criteria;
+pub mod explicit_3criteria;
+pub mod explicit_5criteria;
+pub mod single_criterion;
