+++ examples/tsp/composing_criteria/single_criterion.rs
